backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
motor
passlib[bcrypt]
pyjwt
bcrypt==4.0.1
yt-dlp
python-dotenv
aiofiles
assemblyai   
openai     
requests
python-jose[cryptography]
pydantic[email]
pymongo[srv]
pymongo
google-generativeai